+++ src/testunits/module_BASE/MultiDimNoisyANDTestSuite.h
@@ -36,7 +36,7 @@ namespace gum_tests {
     void testCreationNoisyOr() {
       gum::LabelizedVariable a("a", "", 2), b("b", "", 2), c("c", "", 2),
         d("d", "", 2);
-      gum::MultiDimNoisyAND< float > p(0.2f);
+      gum::MultiDimNoisyAND< double > p(0.2f);
 
       // trying to change weight for a non cause
       TS_ASSERT_THROWS(p.causalWeight(b, 0.4f), gum::InvalidArgument);
@@ -55,13 +55,13 @@ namespace gum_tests {
                        "a<0,1>=noisyAND([0.2],b<0,1>[0.4]c<0,1>[1]d<0,1>[0.7])");
       TS_ASSERT_EQUALS(p.realSize(), (gum::Size)4);
 
-      gum::MultiDimNoisyAND< float > q(p);
+      gum::MultiDimNoisyAND< double > q(p);
       TS_ASSERT_EQUALS(q.toString(),
                        "a<0,1>=noisyAND([0.2],b<0,1>[0.4]c<0,1>[1]d<0,1>[0.7])");
       TS_ASSERT_EQUALS(p.realSize(), (gum::Size)4);
 
       // trying a noisyAND with 0 as external weight (not allowed)
-      TS_ASSERT_THROWS(gum::MultiDimNoisyAND< float > qq(0.0f),
+      TS_ASSERT_THROWS(gum::MultiDimNoisyAND< double > qq(0.0f),
                        gum::InvalidArgument);
     }
 
@@ -71,7 +71,7 @@ namespace gum_tests {
       gum::LabelizedVariable malaria("Malaria", "", 2);
       gum::LabelizedVariable fever("Fever", "", 2);
 
-      gum::MultiDimNoisyAND< float > p(1.0f);
+      gum::MultiDimNoisyAND< double > p(1.0f);
       p << fever << malaria << flu << cold;
       p.causalWeight(cold, 1.0f);
       p.causalWeight(flu, 1.0f);
@@ -108,7 +108,7 @@ namespace gum_tests {
       gum::LabelizedVariable malaria("Malaria", "", 2);
       gum::LabelizedVariable fever("Fever", "", 2);
 
-      gum::MultiDimNoisyAND< float > p(1.0f);
+      gum::MultiDimNoisyAND< double > p(1.0f);
       p << fever << malaria << flu
         << cold;   // malaria, flu and cold are causes of fever
       p.causalWeight(cold, 0.4f);
@@ -139,7 +139,7 @@ namespace gum_tests {
         TS_ASSERT_DELTA(p[i], witness[j], 1e-6);
       }
 
-      gum::MultiDimNoisyAND< float > q(p);
+      gum::MultiDimNoisyAND< double > q(p);
 
       j = 0;
 
@@ -156,7 +156,7 @@ namespace gum_tests {
       gum::LabelizedVariable competition("competition", "", 2);
       gum::LabelizedVariable unemployment("unemployment", "", 2);
 
-      gum::MultiDimNoisyAND< float > p(1 - 0.0001f);
+      gum::MultiDimNoisyAND< double > p(1 - 0.0001f);
       p << unemployment << competition << requirement << motivation << degree
         << lazy;
       p.causalWeight(lazy, 0.8f);
@@ -187,7 +187,7 @@ namespace gum_tests {
         TS_ASSERT_DELTA(p[i], witness[j], 1e-6);
       }
 
-      gum::MultiDimNoisyAND< float > q(p);
+      gum::MultiDimNoisyAND< double > q(p);
 
       j = 0;
