+++ src/testunits/module_BASE/MultiDimNoisyORCompoundTestSuite.h
@@ -34,7 +34,7 @@ namespace gum_tests {
     void testCreationNoisyOr() {
       gum::LabelizedVariable a("a", "", 2), b("b", "", 2), c("c", "", 2),
         d("d", "", 2);
-      gum::MultiDimNoisyORCompound< float > p(0.2f);
+      gum::MultiDimNoisyORCompound< double > p(0.2f);
 
       // trying to change weight for a non cause
       TS_ASSERT_THROWS(p.causalWeight(b, 0.4f), gum::InvalidArgument);
@@ -54,7 +54,7 @@ namespace gum_tests {
         "a<0,1>=noisyORCompound([0.2],b<0,1>[0.4]c<0,1>[1]d<0,1>[0.7])");
       TS_ASSERT_EQUALS(p.realSize(), (gum::Size)4);
 
-      gum::MultiDimNoisyORCompound< float > q(p);
+      gum::MultiDimNoisyORCompound< double > q(p);
       TS_ASSERT_EQUALS(
         q.toString(),
         "a<0,1>=noisyORCompound([0.2],b<0,1>[0.4]c<0,1>[1]d<0,1>[0.7])");
@@ -67,7 +67,7 @@ namespace gum_tests {
       gum::LabelizedVariable malaria("Malaria", "", 2);
       gum::LabelizedVariable fever("Fever", "", 2);
 
-      gum::MultiDimNoisyORCompound< float > p(0.0f);
+      gum::MultiDimNoisyORCompound< double > p(0.0f);
       p << fever << malaria << flu << cold;
       p.causalWeight(cold, 1.0f);
       p.causalWeight(flu, 1.0f);
@@ -104,7 +104,7 @@ namespace gum_tests {
       gum::LabelizedVariable malaria("Malaria", "", 2);
       gum::LabelizedVariable fever("Fever", "", 2);
 
-      gum::MultiDimNoisyORCompound< float > p(0.0f);
+      gum::MultiDimNoisyORCompound< double > p(0.0f);
       p << fever << malaria << flu << cold;
       p.causalWeight(cold, 0.4f);
       p.causalWeight(flu, 0.8f);
@@ -134,7 +134,7 @@ namespace gum_tests {
         TS_ASSERT_DELTA(p[i], witness[j], 1e-6);
       }
 
-      gum::MultiDimNoisyORCompound< float > q(p);
+      gum::MultiDimNoisyORCompound< double > q(p);
 
       j = 0;
 
@@ -151,7 +151,7 @@ namespace gum_tests {
       gum::LabelizedVariable competition("competition", "", 2);
       gum::LabelizedVariable unemployment("unemployment", "", 2);
 
-      gum::MultiDimNoisyORCompound< float > p(0.0001f);
+      gum::MultiDimNoisyORCompound< double > p(0.0001f);
       p << unemployment << competition << requirement << motivation << degree
         << lazy;
       p.causalWeight(lazy, 0.1f);
@@ -180,7 +180,7 @@ namespace gum_tests {
         TS_ASSERT_DELTA(p[i], witness[j], 1e-6);
       }
 
-      gum::MultiDimNoisyORCompound< float > q(p);
+      gum::MultiDimNoisyORCompound< double > q(p);
 
       j = 0;
